show banner in team settings
