+++ packages/app/src/app/overmind/modals.ts
@@ -7,7 +7,6 @@
   actions.modals.myModal.close() (Close it, with optional value matching its result)
 */
 
-import { AlertModalComponents } from 'app/pages/common/Modals/GenericAlertModal';
 import { TeamStep } from 'app/pages/Dashboard/Components/NewTeamModal';
 
 export const forkFrozenModal = {
@@ -45,7 +44,6 @@ export const alertModal: {
     title: string;
     message?: string;
     type?: 'link' | 'primary' | 'danger' | 'secondary';
-    customComponent?: keyof typeof AlertModalComponents;
   };
   result: boolean;
 } = {
