+++ packages/app/src/app/pages/common/Modals/GenericAlertModal/index.tsx
@@ -2,32 +2,10 @@ import React from 'react';
 import { useActions, useAppState } from 'app/overmind';
 import Modal from 'app/components/Modal';
 import { Alert } from '../Common/Alert';
-import { MemberPaymentConfirmation } from '../MemberPaymentConfirmation';
-
-export const AlertModalComponents = { MemberPaymentConfirmation };
 
 export const GenericAlertModal = () => {
   const { alertModal } = useActions().modals;
-  const {
-    title,
-    message,
-    type,
-    isCurrent,
-    customComponent,
-  } = useAppState().modals.alertModal;
-
-  if (customComponent) {
-    const Component = AlertModalComponents[customComponent];
-    return (
-      <Modal
-        isOpen={isCurrent}
-        width={450}
-        onClose={() => alertModal.close(false)}
-      >
-        <Component title={title} />
-      </Modal>
-    );
-  }
+  const { title, message, type, isCurrent } = useAppState().modals.alertModal;
 
   return (
     <Modal
